Affichage preview des CSV côté UI
